refactor(FarePolicy): 요금 정책을 Enum으로 구현
- FareCalculator 삭제

+++ src/main/java/wooteco/subway/domain/PathCalculator.java
@@ -47,7 +47,7 @@ public class PathCalculator {
         return dijkstraShortestPath.getPath(source, target).getVertexList();
     }
 
-    public double calculateShortestDistance(Station source, Station target) {
-        return dijkstraShortestPath.getPathWeight(source, target);
+    public int calculateShortestDistance(Station source, Station target) {
+        return (int) dijkstraShortestPath.getPathWeight(source, target);
     }
 }
